added Spring Boot Security, application.properties redistributed on .yml profiles, added registration/authorization system through JWT tokens.

+++ src/main/java/carsharing/service/CustomerService.java
@@ -2,16 +2,27 @@ package carsharing.service;
 
 import carsharing.dao.model.Customer;
 import carsharing.dao.repository.CustomerRepository;
+import carsharing.dao.repository.RoleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
+import java.util.Collections;
 
 @Service
-public class CustomerService {
+public class CustomerService implements UserDetailsService {
 
     @Autowired
     private CustomerRepository customerRepository;
+    @Autowired
+    private RoleRepository roleRepository;
+    @Autowired
+    public BCryptPasswordEncoder passwordEncoder;
+
 
     public ArrayList<Customer> getAll() {
         return (ArrayList<Customer>) customerRepository.findAll();
@@ -21,15 +32,36 @@ public class CustomerService {
         return customerRepository.findById(id);
     }
 
-    public void delete(Long id) {
-        customerRepository.deleteById(id);
+    public boolean delete(Long id) {
+        if (customerRepository.findById(id) != null) {
+            customerRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
-    public void create(Customer entity) {
-        customerRepository.create(entity);
+    public boolean save(Customer customer) {
+        customer.setRoles(Collections.singleton(roleRepository.findRoleByName("ROLE_USER")));
+        customer.setPassword(passwordEncoder.encode(customer.getPassword()));
+        customerRepository.create(customer);
+        return true;
     }
 
     public void update(Customer entity) {
         customerRepository.update(entity);
+    }
+
+    public Customer getCustomerByEmail(String email) {
+        return customerRepository.getCustomerByEmail(email);
+    }
+
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        Customer customer = getCustomerByEmail(username);
+        if(customer == null) {
+            throw new UsernameNotFoundException("Customer not found");
+        }
+        return customer;
     }
 }
